+++ src/components/EmptyContent.tsx
@@ -7,14 +7,22 @@ import todoEmpty from '@/assets/images/todo_empty.png';
 interface EmptyContentProps {
   mode?: 'todo' | 'done';
   className?: string;
-  children?: React.ReactNode;
 }
 
-const EmptyContent = ({ mode = 'todo', children, className }: EmptyContentProps) => {
+const EmptyContent = ({ mode = 'todo', className }: EmptyContentProps) => {
+  const getMessage = () => {
+    const message = {
+      todo: '할 일이 없어요.\nTODO를 새롭게 추가해주세요',
+      done: '아직 다 한 일이 없어요.\n해야 할 일을 체크해보세요!',
+    };
+    return message[mode];
+  };
   return (
     <div className={clsx('flex flex-col items-center', className)}>
       <Image src={mode === 'todo' ? todoEmpty : doneEmpty} alt={`${mode}초기이미지`} />
-      <div className='text-slate-400 text-center font-bold text-base'>{children}</div>
+      <div className='text-slate-400 text-center font-bold text-base whitespace-pre-line'>
+        {getMessage()}
+      </div>
     </div>
   );
 };
